moved button groups to be with controllers
